Reorganized legacy Region performance testing implementations, and the performance test files for PolygonArea
