+++ GoRogue.PerformanceTests/Regions/MocksAndImplementations/RegionMock.cs
@@ -4,7 +4,7 @@ using JetBrains.Annotations;
 using SadRogue.Primitives;
 using SadRogue.Primitives.PointHashers;
 
-namespace GoRogue.PerformanceTests.Regions
+namespace GoRogue.PerformanceTests.Regions.MocksAndImplementations
 {
     // Object representing a region's internal data and the general method of construction.  Necessary to allow the
     // benchmark functions to actually return a value which depends on the creation method, to avoid any of it being
